Create age rating filter component
